requirements.txt
streamlit
langchain
langchain-google-genai
pillow
python-dotenv
faiss-cpu
numpy
google-generativeai
